LayoutTests/platform/ios/accessibility/visible-character-range-width-changes-expected.txt

This tests that visibleCharacterRange returns expected visible ranges with various view width values.


Testing view width values 100 to 1300.
Range with view 100x500, scrollTop 500: {224, 359}
Range with view 200x500, scrollTop 500: {874, 839}
Range with view 300x500, scrollTop 500: {1249, 1199}
Range with view 400x500, scrollTop 500: {1749, 1679}
Range with view 500x500, scrollTop 500: {2249, 2159}
Range with view 600x500, scrollTop 500: {2749, 2639}
Range with view 700x500, scrollTop 500: {2729, 3009}
Range with view 800x500, scrollTop 500: {2249, 3489}
Range with view 900x500, scrollTop 500: {1814, 3924}
Range with view 1000x500, scrollTop 500: {1479, 4259}
Range with view 1100x500, scrollTop 500: {819, 4919}
Range with view 1200x500, scrollTop 500: {449, 5289}
Range with view 1300x500, scrollTop 500: {0, 5739}

PASS successfullyParsed is true

TEST COMPLETE
